pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sense-music"
version = "0.1.0"
description = "Turn audio into structured analysis and annotated visualizations for AI perception"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Humanjava Enterprises", email = "hello@humanjava.com" },
]
keywords = ["audio-analysis", "music", "spectrogram", "ai", "librosa", "whisper"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
]
dependencies = [
    "librosa>=0.10",
    "matplotlib>=3.7",
    "Pillow>=10.0",
    "numpy",
    "openai-whisper",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "build", "twine"]

[project.urls]
Homepage = "https://huje.tools"
Repository = "https://github.com/HumanjavaEnterprises/huje.sensemusic.OC-python.src"
Documentation = "https://huje.tools"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
